+++ CHANGELOG.md
@@ -58,6 +58,7 @@ and yarn are now dependencies for building the backend.
 - Set default log level to 'warn'.
 - Optimize check marshaling.
 - Silenced API only accepts 'id' parameter on DELETE requests.
+- Disable gostatsd internal metric collection.
 
 ### Fixed
 - Terminate processes gracefully in e2e tests, allowing ports to be reused.